pyproject.toml
[project]
name = "py-aozora-data"
version = "2.1.0"
description = "Aozora-bunko data model"
authors = [{ name = "Kenichi Sato", email = "ksato9700@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "python-dotenv>=1.1",
    "requests>=2.32",
    "boto3>=1.35",
    "algoliasearch",
    "google-auth>=2.0",
]

[project.urls]
Homepage = "https://github.com/your-username/py-aozora-data" # Replace with your project's URL
"Bug Tracker" = "https://github.com/your-username/py-aozora-data/issues" # Replace with your project's issue tracker

[project.scripts]
# If you have any scripts, define them here, e.g.:
# my-script = "aozora_data.cli:main"
sjis-to-utf8 = "aozora_data.sjis_to_utf8.cli:main"
download-all = "aozora_data.download_all:main"
convert-all = "aozora_data.convert_all:main"
text-to-html = "aozora_data.text_to_html.cli:main"
html-convert-all = "aozora_data.html_convert_all:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "requests-mock>=1.12",
    "pre-commit>=4.2",
    "mypy>=1.15",
    "ruff>=0.14.10",
]

[tool.ruff]
# Enable a selection of linter rules
# B: Bugbear
# E: pycodestyle errors
# F: Pyflakes
# I: isort (formatting and sorting imports)
# N: pep8-naming
# W: pycodestyle warnings
# D: pydocstyle (for docstring conventions - optional, remove if not using)
# UP: pyupgrade (for suggesting modern Python syntax)
# ANN: flake8-annotations (for type hints - optional)
# C4: flake8-comprehensions
# SIM: flake8-simplify
# RUF: Ruff-specific rules
lint.select = ["E", "F", "W", "I", "C", "N", "D", "B", "UP", "ANN", "C4", "SIM", "RUF"]
lint.ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in public method
    "D203", # 1 blank line required before docstring
    "D211", # Checks for docstrings on class definitions that are preceded by a blank line.
    "D213", # Checks for docstring summary lines that are not positioned on the second physical line of the docstring.
    "ANN201", # Missing type annotation for function
]
line-length = 102 # Matches Black's default line length for compatibility

[tool.ruff.lint.per-file-ignores]
# Example: Ignore specific rules in a test file
"tests/*" = ["D"] # Ignore docstring rules in test files

[tool.ruff.lint.isort]
# Ruff's built-in isort functionality can be configured here
known-first-party = ["aozora_data"] # Add your project's top-level package name

[tool.uv]
# You typically don't need a specific tool.uv table for dependencies
# as they are now handled by the [project] table for regular and
# [project.optional-dependencies] for dev.
# However, you might use this for uv-specific configurations if they arise in the future.

[tool.mypy]
files = "**/*.py"
plugins = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aozora_data"]
